backend/requirements.txt
flask
flask-cors
pymongo
werkzeug
PyJWT
Pillow
